+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # changelog
 
 ## [unreleased]
+* added `PartialEq` and `Eq` impl for `Stopwatch`
+* updated documentation for `Stopwatch::from_raw`
 
 ## [1.4.0] - 2022-08-20
 * added `Guard::elapsed`